Block library: Extracte deprecated fields to their own files
