+++ packages/block-library/src/image/index.js
@@ -1,24 +1,19 @@
 /**
- * External dependencies
- */
-import classnames from 'classnames';
-
-/**
  * WordPress dependencies
  */
-import { RichText } from '@wordpress/block-editor';
 import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
+import deprecated from './deprecated';
 import edit from './edit';
 import icon from './icon';
 import metadata from './block.json';
 import save from './save';
 import transforms from './transforms';
 
-const { name, attributes: blockAttributes } = metadata;
+const { name } = metadata;
 
 export { metadata, name };
 
@@ -39,80 +34,5 @@ export const settings = {
 	},
 	edit,
 	save,
-	deprecated: [
-		{
-			attributes: blockAttributes,
-			save( { attributes } ) {
-				const { url, alt, caption, align, href, width, height, id } = attributes;
-
-				const classes = classnames( {
-					[ `align${ align }` ]: align,
-					'is-resized': width || height,
-				} );
-
-				const image = (
-					<img
-						src={ url }
-						alt={ alt }
-						className={ id ? `wp-image-${ id }` : null }
-						width={ width }
-						height={ height }
-					/>
-				);
-
-				return (
-					<figure className={ classes }>
-						{ href ? <a href={ href }>{ image }</a> : image }
-						{ ! RichText.isEmpty( caption ) && <RichText.Content tagName="figcaption" value={ caption } /> }
-					</figure>
-				);
-			},
-		},
-		{
-			attributes: blockAttributes,
-			save( { attributes } ) {
-				const { url, alt, caption, align, href, width, height, id } = attributes;
-
-				const image = (
-					<img
-						src={ url }
-						alt={ alt }
-						className={ id ? `wp-image-${ id }` : null }
-						width={ width }
-						height={ height }
-					/>
-				);
-
-				return (
-					<figure className={ align ? `align${ align }` : null } >
-						{ href ? <a href={ href }>{ image }</a> : image }
-						{ ! RichText.isEmpty( caption ) && <RichText.Content tagName="figcaption" value={ caption } /> }
-					</figure>
-				);
-			},
-		},
-		{
-			attributes: blockAttributes,
-			save( { attributes } ) {
-				const { url, alt, caption, align, href, width, height } = attributes;
-				const extraImageProps = width || height ? { width, height } : {};
-				const image = <img src={ url } alt={ alt } { ...extraImageProps } />;
-
-				let figureStyle = {};
-
-				if ( width ) {
-					figureStyle = { width };
-				} else if ( align === 'left' || align === 'right' ) {
-					figureStyle = { maxWidth: '50%' };
-				}
-
-				return (
-					<figure className={ align ? `align${ align }` : null } style={ figureStyle }>
-						{ href ? <a href={ href }>{ image }</a> : image }
-						{ ! RichText.isEmpty( caption ) && <RichText.Content tagName="figcaption" value={ caption } /> }
-					</figure>
-				);
-			},
-		},
-	],
+	deprecated,
 };
