+++ packages/block-library/src/list/index.js
@@ -1,71 +1,32 @@
 /**
- * External dependencies
- */
-import { omit } from 'lodash';
-
-/**
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { RichText } from '@wordpress/block-editor';
 
 /**
  * Internal dependencies
  */
+import deprecated from './deprecated';
 import edit from './edit';
 import icon from './icon';
 import metadata from './block.json';
 import save from './save';
 import transforms from './transforms';
 
-const { name, attributes: schema } = metadata;
+const { name } = metadata;
 
 export { metadata, name };
-
-const supports = {
-	className: false,
-};
 
 export const settings = {
 	title: __( 'List' ),
 	description: __( 'Create a bulleted or numbered list.' ),
 	icon,
 	keywords: [ __( 'bullet list' ), __( 'ordered list' ), __( 'numbered list' ) ],
-	supports,
+	supports: {
+		className: false,
+	},
 	transforms,
-	deprecated: [
-		{
-			supports,
-			attributes: {
-				...omit( schema, [ 'ordered' ] ),
-				nodeName: {
-					type: 'string',
-					source: 'property',
-					selector: 'ol,ul',
-					property: 'nodeName',
-					default: 'UL',
-				},
-			},
-			migrate( attributes ) {
-				const { nodeName, ...migratedAttributes } = attributes;
-
-				return {
-					...migratedAttributes,
-					ordered: 'OL' === nodeName,
-				};
-			},
-			save( { attributes } ) {
-				const { nodeName, values } = attributes;
-
-				return (
-					<RichText.Content
-						tagName={ nodeName.toLowerCase() }
-						value={ values }
-					/>
-				);
-			},
-		},
-	],
+	deprecated,
 	merge( attributes, attributesToMerge ) {
 		const { values } = attributesToMerge;
